+++ geopackage-sdk/src/main/java/mil/nga/geopackage/features/OAPIFeatureGenerator.java
@@ -13,11 +13,11 @@ import mil.nga.geopackage.features.user.FeatureRow;
 import mil.nga.sf.Geometry;
 
 /**
- * OGC OpenAPI Feature Generator
+ * OGC API Feature Generator
  *
  * @author osbornb
  */
-public class OpenAPIFeatureGenerator extends OpenAPIFeatureCoreGenerator {
+public class OAPIFeatureGenerator extends OAPIFeatureCoreGenerator {
 
     /**
      * Feature DAO
@@ -32,8 +32,8 @@ public class OpenAPIFeatureGenerator extends OpenAPIFeatureCoreGenerator {
      * @param server     server url
      * @param name       collection identifier
      */
-    public OpenAPIFeatureGenerator(GeoPackage geoPackage, String tableName,
-                                   String server, String name) {
+    public OAPIFeatureGenerator(GeoPackage geoPackage, String tableName,
+                                String server, String name) {
         super(geoPackage, tableName, server, name);
     }